refactor subtitle reading
We can now detect the character encoding of a subtitle file, and the
language. The first is used to translate non-utf8 SRT files to
utf-8 vtt files. The second is used to detect the language of a SRT
or VTT file that doesn't have the language encoded as a tag in the
filename.